moving uploading to batch_upload
